fix(SC-01): extract 6 layouts to server component + client shell pattern
- receptionist: extract to receptionist-layout-shell.tsx
- super-admin: extract to super-admin-layout-shell.tsx
- equipment: extract to equipment-layout-shell.tsx (preserves i18n context)
- specialist: extract to specialist-layout-shell.tsx
- lab: extract to lab-layout-shell.tsx
- pharmacist: extract to pharmacist-layout-shell.tsx

All layout.tsx files are now server components that delegate to
'use client' shell components, matching the doctor/admin/patient pattern.
