+++ src/app/(specialist)/layout.tsx
@@ -1,10 +1,6 @@
-"use client";
-
-import { usePathname } from "next/navigation";
-import {
-  ClinicDashboardLayout,
-} from "@/components/layouts/clinic-dashboard-layout";
-import { getSpecialistConfigFromPathname } from "@/config/specialist-registry";
+// SC-01: Server component layout — client-interactive shell extracted to
+// a separate "use client" component (specialist-layout-shell).
+import { SpecialistLayoutShell } from "@/components/layouts/specialist-layout-shell";
 
 /**
  * Consolidated layout for all specialist dashboard route groups.
@@ -21,22 +17,5 @@ export default function SpecialistLayout({
 }: {
   children: React.ReactNode;
 }) {
-  const pathname = usePathname();
-
-  // Equipment has its own nested layout — pass children through
-  if (pathname.startsWith("/equipment")) {
-    return <>{children}</>;
-  }
-
-  const config = getSpecialistConfigFromPathname(pathname);
-
-  if (!config) {
-    return <>{children}</>;
-  }
-
-  return (
-    <ClinicDashboardLayout config={config}>
-      {children}
-    </ClinicDashboardLayout>
-  );
+  return <SpecialistLayoutShell>{children}</SpecialistLayoutShell>;
 }
